requirements.txt
# Core dependencies
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# Solana dependencies
solana==0.32.0
solders==0.18.1
anchorpy==0.18.0
spl==0.2.0
solana-py==0.32.0

# Async and networking
aiohttp==3.9.1
asyncio-mqtt==0.16.1
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Trading and analysis
pandas==2.1.4
numpy==1.25.2
ccxt==4.1.77
ta==0.10.2

# Security and encryption
cryptography==41.0.8
pyjwt==2.8.0
base58==2.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API and web
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# Solana specific
jupiter-python-sdk==0.1.0
raydium-sdk==0.1.0
serum-python==0.1.0

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1 